refactor(demod): move HannBank / HANN_LENS / window_idx_for_snr* from snr to crate::demod (#85 B1)
